requirements.txt
Flask==3.0.3
mysql-connector-python==8.4.0
python-dotenv==1.0.1
bcrypt==4.1.3
Flask-Session==0.8.0
Werkzeug==3.0.3
